ModelReader implementation
