+++ deeplearning4j-scaleout/deeplearning4j-nlp/src/main/java/org/deeplearning4j/models/embeddings/inmemory/JointStorage.java
@@ -284,6 +284,11 @@ public class JointStorage<T extends SequenceElement> implements WeightLookupTabl
         return 0;
     }
 
+    @Override
+    public VocabCache<T> getVocabCache() {
+        return this;
+    }
+
     public static class Builder<T extends SequenceElement> {
         private Map<Long, WeightLookupTable<T>> mapTables = new ConcurrentHashMap<>();
         private Map<Long, VocabCache<T>> mapVocabs = new ConcurrentHashMap<>();
