+++ deeplearning4j-scaleout/deeplearning4j-nlp/src/main/java/org/deeplearning4j/models/embeddings/wordvectors/WordVectorsImpl.java
@@ -18,21 +18,16 @@
 
 package org.deeplearning4j.models.embeddings.wordvectors;
 
-import com.google.common.collect.Lists;
 import com.google.common.util.concurrent.AtomicDouble;
 import lombok.Getter;
-import org.deeplearning4j.berkeley.Counter;
-import org.deeplearning4j.clustering.sptree.DataPoint;
 import org.deeplearning4j.clustering.vptree.VPTree;
+import org.deeplearning4j.models.embeddings.reader.ModelUtils;
+import org.deeplearning4j.models.embeddings.reader.impl.BasicModelUtils;
 import org.deeplearning4j.models.sequencevectors.sequence.SequenceElement;
 import org.deeplearning4j.models.embeddings.WeightLookupTable;
-import org.deeplearning4j.models.embeddings.inmemory.InMemoryLookupTable;
 import org.deeplearning4j.models.word2vec.wordstore.VocabCache;
-import org.deeplearning4j.util.MathUtils;
-import org.deeplearning4j.util.SetUtils;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.factory.Nd4j;
-import org.nd4j.linalg.ops.transforms.Transforms;
 
 import java.util.*;
 
@@ -47,6 +42,7 @@ public class WordVectorsImpl<T extends SequenceElement> implements WordVectors {
     @Getter protected WeightLookupTable<T> lookupTable;
     @Getter protected VocabCache<T> vocab;
     @Getter protected int layerSize = 100;
+    @Getter protected ModelUtils<T> modelUtils;
 
     protected int numIterations = 1;
     protected int numEpochs = 1;
@@ -83,66 +79,9 @@ public class WordVectorsImpl<T extends SequenceElement> implements WordVectors {
      * @return the words nearest the mean of the words
      */
     public Collection<String> wordsNearestSum(Collection<String> positive,Collection<String> negative,int top) {
-        INDArray words = Nd4j.create(lookupTable().layerSize());
-        Set<String> union = SetUtils.union(new HashSet<>(positive), new HashSet<>(negative));
-        for(String s : positive)
-            words.addi(lookupTable().vector(s));
-
-
-        for(String s : negative)
-            words.addi(lookupTable.vector(s).mul(-1));
-
-        if(lookupTable() instanceof InMemoryLookupTable) {
-            InMemoryLookupTable l = (InMemoryLookupTable) lookupTable();
-            INDArray syn0 = l.getSyn0();
-            INDArray weights = syn0.norm2(0).rdivi(1).muli(words);
-            INDArray distances = syn0.mulRowVector(weights).sum(1);
-            INDArray[] sorted = Nd4j.sortWithIndices(distances,0,false);
-            INDArray sort = sorted[0];
-            List<String> ret = new ArrayList<>();
-            if(top > sort.length())
-                top = sort.length();
-            //there will be a redundant word
-            int end = top;
-            for(int i = 0; i < end; i++) {
-                String word = vocab.wordAtIndex(sort.getInt(i));
-                if(union.contains(word)) {
-                    end++;
-                    if(end >= sort.length())
-                        break;
-                    continue;
-                }
-
-                String add = vocab().wordAtIndex(sort.getInt(i));
-                if(add == null || add.equals("UNK") || add.equals("STOP")) {
-                    end++;
-                    if(end >= sort.length())
-                        break;
-                    continue;
-                }
-
-
-                ret.add(vocab().wordAtIndex(sort.getInt(i)));
-            }
-
-
-            return ret;
-        }
-
-        Counter<String> distances = new Counter<>();
-
-        for(String s : vocab().words()) {
-            INDArray otherVec = getWordVectorMatrix(s);
-            double sim = Transforms.cosineSim(words, otherVec);
-            distances.incrementCount(s, sim);
-        }
-
-
-        distances.keepTopNKeys(top);
-        return distances.keySet();
-
-
+        return modelUtils.wordsNearestSum(positive, negative, top);
     }
+
     /**
      * Words nearest based on positive and negative words
      * * @param top the top n words
@@ -150,50 +89,9 @@ public class WordVectorsImpl<T extends SequenceElement> implements WordVectors {
      */
     @Override
     public Collection<String> wordsNearestSum(INDArray words,int top) {
-
-        if(lookupTable() instanceof InMemoryLookupTable) {
-            InMemoryLookupTable l = (InMemoryLookupTable) lookupTable();
-            INDArray syn0 = l.getSyn0();
-            INDArray weights = syn0.norm2(0).rdivi(1).muli(words);
-            INDArray distances = syn0.mulRowVector(weights).sum(1);
-            INDArray[] sorted = Nd4j.sortWithIndices(distances,0,false);
-            INDArray sort = sorted[0];
-            List<String> ret = new ArrayList<>();
-            if(top > sort.length())
-                top = sort.length();
-            //there will be a redundant word
-            int end = top;
-            for(int i = 0; i < end; i++) {
-                String add = vocab().wordAtIndex(sort.getInt(i));
-                if(add == null || add.equals("UNK") || add.equals("STOP")) {
-                    end++;
-                    if(end >= sort.length())
-                        break;
-                    continue;
-                }
-
-
-                ret.add(vocab().wordAtIndex(sort.getInt(i)));
-            }
-
-
-            return ret;
-        }
-
-        Counter<String> distances = new Counter<>();
-
-        for(String s : vocab().words()) {
-            INDArray otherVec = getWordVectorMatrix(s);
-            double sim = Transforms.cosineSim(words, otherVec);
-            distances.incrementCount(s, sim);
-        }
-
-
-        distances.keepTopNKeys(top);
-        return distances.keySet();
-
-
+        return modelUtils.wordsNearestSum(words, top);
     }
+
     /**
      * Words nearest based on positive and negative words
      * * @param top the top n words
@@ -201,50 +99,7 @@ public class WordVectorsImpl<T extends SequenceElement> implements WordVectors {
      */
     @Override
     public Collection<String> wordsNearest(INDArray words, int top) {
-        if(lookupTable() instanceof InMemoryLookupTable) {
-            InMemoryLookupTable l = (InMemoryLookupTable) lookupTable();
-            INDArray syn0 = l.getSyn0();
-            INDArray weights = syn0.norm2(0).rdivi(1).muli(words);
-            INDArray distances = syn0.mulRowVector(weights).mean(1);
-            INDArray[] sorted = Nd4j.sortWithIndices(distances,0,false);
-            INDArray sort = sorted[0];
-            List<String> ret = new ArrayList<>();
-            if(top > sort.length())
-                top = sort.length();
-            //there will be a redundant word
-            int end = top;
-            for(int i = 0; i < end; i++) {
-                VocabCache vocabCache = vocab();
-                int s = sort.getInt(0, i);
-                String add = vocabCache.wordAtIndex(s);
-                if(add == null || add.equals("UNK") || add.equals("STOP")) {
-                    end++;
-                    if(end >= sort.length())
-                        break;
-                    continue;
-                }
-
-
-                ret.add(vocabCache.wordAtIndex(s));
-            }
-
-
-            return ret;
-        }
-
-        Counter<String> distances = new Counter<>();
-
-        for(String s : vocab().words()) {
-            INDArray otherVec = getWordVectorMatrix(s);
-            double sim = Transforms.cosineSim(words, otherVec);
-            distances.incrementCount(s, sim);
-        }
-
-
-        distances.keepTopNKeys(top);
-        return distances.keySet();
-
-
+        return modelUtils.wordsNearest(words, top);
     }
 
     /**
@@ -254,101 +109,19 @@ public class WordVectorsImpl<T extends SequenceElement> implements WordVectors {
      * @return the top n words
      */
     public Collection<String> wordsNearestSum(String word,int n) {
-        INDArray vec = Transforms.unitVec(this.getWordVectorMatrix(word));
-
-
-        if(lookupTable() instanceof InMemoryLookupTable) {
-            InMemoryLookupTable l = (InMemoryLookupTable) lookupTable();
-            INDArray syn0 = l.getSyn0();
-            INDArray weights = syn0.norm2(0).rdivi(1).muli(vec);
-            INDArray distances = syn0.mulRowVector(weights).sum(1);
-            INDArray[] sorted = Nd4j.sortWithIndices(distances,0,false);
-            INDArray sort = sorted[0];
-            List<String> ret = new ArrayList<>();
-            SequenceElement word2 = vocab().wordFor(word);
-            if(n > sort.length())
-                n = sort.length();
-            //there will be a redundant word
-            for(int i = 0; i < n + 1; i++) {
-                if(sort.getInt(i) == word2.getIndex())
-                    continue;
-                String add = vocab().wordAtIndex(sort.getInt(i));
-                if(add == null || add.equals("UNK") || add.equals("STOP")) {
-                    continue;
-                }
-
-
-
-                ret.add(vocab().wordAtIndex(sort.getInt(i)));
-            }
-
-
-            return ret;
-        }
-
-        if(vec == null)
-            return new ArrayList<>();
-        Counter<String> distances = new Counter<>();
-
-        for(String s : vocab().words()) {
-            if(s.equals(word))
-                continue;
-            INDArray otherVec = getWordVectorMatrix(s);
-            double sim = Transforms.cosineSim(vec,otherVec);
-            distances.incrementCount(s, sim);
-        }
-
-
-        distances.keepTopNKeys(n);
-        return distances.keySet();
-
+        return modelUtils.wordsNearestSum(word, n);
     }
 
 
-    /** Accuracy based on questions which are a space separated list of strings
+    /**
+     * Accuracy based on questions which are a space separated list of strings
     * where the first word is the query word, the next 2 words are negative,
-            * and the last word is the predicted word to be nearest
+    * and the last word is the predicted word to be nearest
     * @param questions the questions to ask
     * @return the accuracy based on these questions
     */
     public Map<String,Double> accuracy(List<String> questions) {
-        Map<String,Double> accuracy = new HashMap<>();
-        Counter<String> right = new Counter<>();
-        String analogyType = "";
-        for(String s : questions) {
-            if(s.startsWith(":")) {
-                double correct = right.getCount("correct");
-                double wrong = right.getCount("wrong");
-                if(analogyType.isEmpty()){
-                    analogyType=s;
-                    continue;
-                }
-                double accuracyRet = 100.0 * correct / (correct + wrong);
-                accuracy.put(analogyType,accuracyRet);
-                analogyType = s;
-                right.clear();
-            }
-            else {
-                String[] split = s.split(" ");
-                String word = split[0];
-                List<String> positive = Arrays.asList(word);
-                List<String> negative = Arrays.asList(split[1],split[2]);
-                String predicted = split[3];
-                String w = wordsNearest(positive,negative,1).iterator().next();
-                if(predicted.equals(w))
-                    right.incrementCount("right",1.0);
-                else
-                    right.incrementCount("wrong",1.0);
-
-            }
-        }
-        if(!analogyType.isEmpty()){
-            double correct = right.getCount("correct");
-            double wrong = right.getCount("wrong");
-            double accuracyRet = 100.0 * correct / (correct + wrong);
-            accuracy.put(analogyType,accuracyRet);
-        }
-        return accuracy;
+        return modelUtils.accuracy(questions);
     }
 
     @Override
@@ -365,12 +138,7 @@ public class WordVectorsImpl<T extends SequenceElement> implements WordVectors {
      * @return the list of words that are similar in the vocab
      */
     public List<String> similarWordsInVocabTo(String word,double accuracy) {
-        List<String> ret = new ArrayList<>();
-        for(String s : vocab.words()) {
-            if(MathUtils.stringSimilarity(word, s) >= accuracy)
-                ret.add(s);
-        }
-        return ret;
+        return this.modelUtils.similarWordsInVocabTo(word, accuracy);
     }
 
     /**
@@ -415,7 +183,7 @@ public class WordVectorsImpl<T extends SequenceElement> implements WordVectors {
      */
     @Override
     public Collection<String> wordsNearest(Collection<String> positive, Collection<String> negative, int top) {
-        return null;
+        return modelUtils.wordsNearest(positive, negative, top);
     }
 
     /**
@@ -425,25 +193,19 @@ public class WordVectorsImpl<T extends SequenceElement> implements WordVectors {
      * @return the top n words
      */
     public Collection<String> wordsNearest(String word,int n) {
-       return null;
+       return modelUtils.wordsNearest(word, n);
     }
 
 
     /**
-     * Returns the similarity of 2 words. Result value will be in range [-1,1], where -1.0 is exact opposite similarity, i.e. NO similarity, and 1.0 is total match of two word vectors.
-     * However, most of time you'll see values in range [0,1], but that's something depends of training corpus.
+     * Returns similarity of two elements, provided by ModelUtils
      *
      * @param word the first word
      * @param word2 the second word
      * @return a normalized similarity (cosine similarity)
      */
     public double similarity(String word,String word2) {
-        if(word.equals(word2))
-            return 1.0;
-
-        if(getWordVectorMatrix(word) == null || getWordVectorMatrix(word2) == null)
-            return -1;
-        return  Transforms.cosineSim(getWordVectorMatrix(word), getWordVectorMatrix(word2));
+        return modelUtils.similarity(word, word2);
     }
 
     @Override
@@ -458,6 +220,9 @@ public class WordVectorsImpl<T extends SequenceElement> implements WordVectors {
 
     public void setLookupTable(WeightLookupTable lookupTable) {
         this.lookupTable = lookupTable;
+        if (modelUtils == null) this.modelUtils = new BasicModelUtils<T>();
+
+        this.modelUtils.init(lookupTable);
     }
 
     public void setVocab(VocabCache vocab) {
